Ref, random chance as double

+++ Random.cpp
@@ -16,7 +16,7 @@ int Random::getInt(int min, int max) {
 	return rand()%(max-min+1)+min;
 }
 
-bool Random::tryChance(int prob) { // prob <0..100>
+bool Random::tryChance(double prob) { // prob <0..100>
 	int l = getInt(0, 100);
 	if (prob > l) return true;
 	else return false;
